improve precision of copper traces

+++ diylc/diylc-core/src/org/diylc/components/AbstractSchematicLeadedSymbol.java
@@ -36,7 +36,7 @@ public abstract class AbstractSchematicLeadedSymbol<T> extends AbstractLeadedCom
   }
 
   @Override
-  protected int getLeadThickness() {
+  protected float getLeadThickness() {
     return 1;
   }
